Fixed ticket #453, double instead of float conversions.

+++ modules/basegl/processors/positionwidgetprocessor.cpp
@@ -94,7 +94,7 @@ void PositionWidgetProcessor::updateWidgetPositionFromPicking(const PickingObjec
     if (move.x == 0.f && move.y == 0.f) return;
 
     vec2 pos = p->getPickingPosition();
-    float depth = p->getPickingDepth();
+    float depth = static_cast<float>(p->getPickingDepth());
     //TODO: do we not need to incorporate transformations here?
     vec3 startNdc = vec3((2.f*pos.x)-1.f, (2.f*pos.y)-1.f, depth);
     vec3 endNdc = vec3((2.f*(pos.x+move.x))-1.f, (2.f*(pos.y+move.y))-1.f, depth);
